FileSec: Restructure
Remove FileSec class and move fileSec generation to METSWriter, like
structMaps. Generate files in a flat hierarchy, instead of nested. Add
grouping based on USE.

Distinguish between a path (for FLocat in fileSec) and a label (for
structMap). Label defaults to the basename of the path. Update test for
this.