Create formatted block comments for Utils
This commit introduces a Django style commenting block for the Utils
module following discussion around the original PR submitted against
this code. The block-comments are significantly different from the
alternative single hash (`# Comment.`) approach which has a different
set of guidelines associated with it in the Artefactual coding
style-guide.

It is hoped the new block comment style will be clearer
to devs looking at the code for the first time.