premisrw: always return unicode in text getter
In metsrw 0.2.4 or older, premisrw was returning:

  - In Py2: binary (bytes)
  - In Py3: text (unicode)

In metsrw 0.3.0, the Py2 code path was updated to explicitly use the `utf-8`
codec since the default is `ascii` which is usually a problem during encoding.

This commit updates the same code path again so in Py2 we also return unicode
making the API more uniform.